+++ test/comptime/default_invalid_float.zig
@@ -6,7 +6,7 @@ comptime {
     const cmd = parsz.Command{
         .name = "test",
         .args = &.{
-            .{ .name = "bad_option", .kind = .option, .value_type = .float, .long = "ratio", .default = "not_a_number" }, // Invalid: not a valid float
+            .{ .name = "bad_option", .kind = .option, .value_type = .f64, .long = "ratio", .default = "not_a_number" }, // Invalid: not a valid float
         },
     };
     _ = parsz.ParseResult(cmd);
